requirements.txt
apache-airflow==3.1.3
requests==2.32.5
python-dotenv==1.2.1
minio==7.2.20